Moved all Adapter Actions in their folder.
