+++ BHoM_Adapter/AdapterActions/Push.cs
@@ -1,0 +1,85 @@
+/*
+ * This file is part of the Buildings and Habitats object Model (BHoM)
+ * Copyright (c) 2015 - 2018, the respective contributors. All rights reserved.
+ *
+ * Each contributor holds copyright over their respective contributions.
+ * The project versioning (Git) records all such contribution source information.
+ *                                           
+ *                                                                              
+ * The BHoM is free software: you can redistribute it and/or modify         
+ * it under the terms of the GNU Lesser General Public License as published by  
+ * the Free Software Foundation, either version 3.0 of the License, or          
+ * (at your option) any later version.                                          
+ *                                                                              
+ * The BHoM is distributed in the hope that it will be useful,              
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of               
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
+ * GNU Lesser General Public License for more details.                          
+ *                                                                            
+ * You should have received a copy of the GNU Lesser General Public License     
+ * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
+ */
+
+using BH.oM.Base;
+using BH.Engine.Base;
+using BH.oM.Data.Requests;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+
+namespace BH.Adapter
+{
+    public abstract partial class BHoMAdapter
+    {
+        /******************************************************/
+        /**** Public Adapter Methods "Adapter ACTIONS"    *****/
+        /******************************************************/
+        /* These methods represent Actions that the Adapter can complete. 
+           They are publicly available in the UI as individual components, e.g. in Grasshopper, under BHoM/Adapters tab. */
+
+        // Performs the full CRUD if implemented, or calls the appropriate basic CRUD/Create method.
+        public virtual List<IObject> Push(IEnumerable<IObject> objects, string tag = "", Dictionary<string, object> pushConfig = null)
+        {
+            bool success = true;
+
+            // Get the Push Type from the pushConfig.
+            string pushType;
+            object ptObj;
+            if (pushConfig != null && pushConfig.TryGetValue("PushType", out ptObj))
+                pushType = ptObj.ToString();
+            else
+                pushType = "Replace";
+
+            // Wrap non-BHoM objects into a Custom BHoMObject to make them work as BHoMObjects.
+            List<IObject> objectsToPush = Modify.WrapNonBHoMObjects(objects, Config, tag, pushConfig).ToList();
+
+            // Clone the objects for immutability in the UI. CloneBeforePush should always be true, except for very specific cases.
+            objectsToPush = Config.CloneBeforePush ? objectsToPush.Select(x => x.DeepClone()).ToList() : objects.ToList();
+
+            // Perform the actual Push.
+            Type iBHoMObjectType = typeof(IBHoMObject);
+            MethodInfo miToList = typeof(Enumerable).GetMethod("Cast");
+            foreach (var typeGroup in objectsToPush.GroupBy(x => x.GetType()))
+            {
+                MethodInfo miListObject = miToList.MakeGenericMethod(new[] { typeGroup.Key });
+
+                var list = miListObject.Invoke(typeGroup, new object[] { typeGroup });
+
+                if (iBHoMObjectType.IsAssignableFrom(typeGroup.Key))
+                {
+                    if (pushType == "Replace")
+                        success &= CRUD(list as dynamic, tag);
+                    else if (pushType == "UpdateOnly")
+                    {
+                        success &= UpdateOnly(list as dynamic, tag);
+                    }
+                }
+            }
+
+            return success ? objectsToPush : new List<IObject>();
+        }
+
+    }
+}
