+++ BHoM_Adapter/AdapterActions/Remove.cs
@@ -1,0 +1,51 @@
+/*
+ * This file is part of the Buildings and Habitats object Model (BHoM)
+ * Copyright (c) 2015 - 2018, the respective contributors. All rights reserved.
+ *
+ * Each contributor holds copyright over their respective contributions.
+ * The project versioning (Git) records all such contribution source information.
+ *                                           
+ *                                                                              
+ * The BHoM is free software: you can redistribute it and/or modify         
+ * it under the terms of the GNU Lesser General Public License as published by  
+ * the Free Software Foundation, either version 3.0 of the License, or          
+ * (at your option) any later version.                                          
+ *                                                                              
+ * The BHoM is distributed in the hope that it will be useful,              
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of               
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
+ * GNU Lesser General Public License for more details.                          
+ *                                                                            
+ * You should have received a copy of the GNU Lesser General Public License     
+ * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
+ */
+
+using BH.oM.Base;
+using BH.Engine.Base;
+using BH.oM.Data.Requests;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+
+namespace BH.Adapter
+{
+    public abstract partial class BHoMAdapter
+    {
+        /******************************************************/
+        /**** Public Adapter Methods "Adapter ACTIONS"    *****/
+        /******************************************************/
+        /* These methods represent Actions that the Adapter can complete. 
+           They are publicly available in the UI as individual components, e.g. in Grasshopper, under BHoM/Adapters tab. */
+
+        // To be renamed to "Remove" or something different from the CRUD method "Delete".
+        // Calls the basic CRUD/Delete method.
+        public virtual int Delete(IRequest request, Dictionary<string, object> config = null)
+        {
+            // If the provided request is a FilterRequest, the specific wrapper method for FilterRequest is called.
+            // For all other cases, Toolkits should implement specific IRequests and the related CRUD Wrapper method(s).
+            return Delete(request as dynamic);
+        }
+    }
+}
