+++ BHoM_Adapter/BHoM_Adapter.csproj
@@ -84,6 +84,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AdapterActions\Execute.cs" />
+    <Compile Include="AdapterActions\Remove.cs" />
+    <Compile Include="AdapterActions\Move.cs" />
+    <Compile Include="AdapterActions\Pull.cs" />
+    <Compile Include="AdapterActions\Push.cs" />
     <Compile Include="AdapterConfig.cs" />
     <Compile Include="CRUD\Create.cs" />
     <Compile Include="CRUD\Delete.cs" />
